requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-ckeditor>=6.0,<7.0
django-redis>=4.12,<5.0
googletrans==4.0.0-rc1